feat: Update branding and theme options
Introduce a new "Flowkh" logo and rename the previous "Kisskh" logo component.
Add a "yellow" theme option to the application's color themes and update the `ThemeSwitcher` component to accommodate this new option with a grid layout.
Update the index.html to include the Montserrat font for potential future use.
Adjust the `TuneIcon` and remove `PaletteIcon` from the Navbar component.